Moved the basic validation rules to a separate namespace to keep the directory more organized.
