+++ validation/rules/EmptyValidationRule.php
@@ -1,17 +1,17 @@
-<?php namespace spitfire\validation;
+<?php namespace spitfire\validation\rules;
 
 /**
- * This filter ensures that a number provided to it was positive.
+ * Validates a value is not empty. This allows the system to make sure that a 
+ * required field has not been field with useless data.
  * 
  * @author César de la Cal <cesar@magic3w.com>
  */
-class PositiveNumberValidationRule implements ValidationRule
+class EmptyValidationRule implements \spitfire\validation\ValidationRule
 {
 	
 	/**
 	 * A message the validation error generated by this object should carry to give
 	 * the end user information about the reason his input was rejected.
-	 * 
 	 * @var string
 	 */
 	private $message;
@@ -19,7 +19,6 @@ class PositiveNumberValidationRule implements ValidationRule
 	/**
 	 * Additional information given to the user in case the validation did not 
 	 * succeed. This message can hold additional infos on how to solve the error.
-	 * 
 	 * @var string
 	 */
 	private $extendedMessage;
@@ -33,12 +32,13 @@ class PositiveNumberValidationRule implements ValidationRule
 	 * Tests a value with this validation rule. Returns the errors detected for
 	 * this element or boolean false on no errors.
 	 * 
+	 * @assert ('ok') == false
 	 * @param mixed $value
 	 * @param mixed $source
 	 * @return \spitfire\validation\ValidationError|boolean
 	 */
 	public function test($value) {
-		if ($value < 0) {
+		if (empty($value)) {
 			return new ValidationError($this->message, $this->extendedMessage);
 		}
 		return false;
