+++ validation/rules/MinLengthValidationRule.php
@@ -1,20 +1,20 @@
-<?php namespace spitfire\validation;
+<?php namespace spitfire\validation\rules;
 
 /**
- * Validates a value is not empty. This allows the system to make sure that a 
- * required field has not been field with useless data.
+ * Validates that the length of a content is bigger than the indicated minimum
+ * length. This validates only data as strings which may cause unexpected 
+ * behavior if you try to test the length of an array.
  * 
  * @author César de la Cal <cesar@magic3w.com>
  */
-class RegexValidationRule implements ValidationRule
+class MinLengthValidationRule implements \spitfire\validation\ValidationRule
 {
-	
 	/**
-	 * The regular expression we wish to test the content of the validator against.
-	 *
-	 * @var string
+	 * The minimum length a string passed to this element must have for the test
+	 * to be passed successfully.
+	 * @var int
 	 */
-	private $regex;
+	private $length;
 	
 	/**
 	 * A message the validation error generated by this object should carry to give
@@ -26,13 +26,12 @@ class RegexValidationRule implements ValidationRule
 	/**
 	 * Additional information given to the user in case the validation did not 
 	 * succeed. This message can hold additional infos on how to solve the error.
-	 * 
 	 * @var string
 	 */
 	private $extendedMessage;
 	
-	public function __construct($regex, $message, $extendedMessage = '') {
-		$this->regex = $regex;
+	public function __construct($length, $message, $extendedMessage = '') {
+		$this->length = $length;
 		$this->message = $message;
 		$this->extendedMessage = $extendedMessage;
 	}
@@ -46,10 +45,10 @@ class RegexValidationRule implements ValidationRule
 	 * @return \spitfire\validation\ValidationError|boolean
 	 */
 	public function test($value) {
-		if (preg_match($this->regex, $value) !== 1) {
+		if (strlen($value) < $this->length) {
 			return new ValidationError($this->message, $this->extendedMessage);
 		}
 		return false;
 	}
-	
+
 }
