+++ validation/rules/PositiveNumberValidationRule.php
@@ -1,17 +1,17 @@
-<?php namespace spitfire\validation;
+<?php namespace spitfire\validation\rules;
 
 /**
- * Validates a value is not empty. This allows the system to make sure that a 
- * required field has not been field with useless data.
+ * This filter ensures that a number provided to it was positive.
  * 
  * @author César de la Cal <cesar@magic3w.com>
  */
-class EmptyValidationRule implements ValidationRule
+class PositiveNumberValidationRule implements \spitfire\validation\ValidationRule
 {
 	
 	/**
 	 * A message the validation error generated by this object should carry to give
 	 * the end user information about the reason his input was rejected.
+	 * 
 	 * @var string
 	 */
 	private $message;
@@ -19,6 +19,7 @@ class EmptyValidationRule implements ValidationRule
 	/**
 	 * Additional information given to the user in case the validation did not 
 	 * succeed. This message can hold additional infos on how to solve the error.
+	 * 
 	 * @var string
 	 */
 	private $extendedMessage;
@@ -32,13 +33,12 @@ class EmptyValidationRule implements ValidationRule
 	 * Tests a value with this validation rule. Returns the errors detected for
 	 * this element or boolean false on no errors.
 	 * 
-	 * @assert ('ok') == false
 	 * @param mixed $value
 	 * @param mixed $source
 	 * @return \spitfire\validation\ValidationError|boolean
 	 */
 	public function test($value) {
-		if (empty($value)) {
+		if ($value < 0) {
 			return new ValidationError($this->message, $this->extendedMessage);
 		}
 		return false;
