+++ validation/rules/RegexValidationRule.php
@@ -1,20 +1,23 @@
-<?php namespace spitfire\validation;
+<?php namespace spitfire\validation\rules;
+
+use spitfire\validation\ValidationError;
+use spitfire\validation\ValidationRule;
 
 /**
- * Validates that the length of a content is bigger than the indicated minimum
- * length. This validates only data as strings which may cause unexpected 
- * behavior if you try to test the length of an array.
+ * Validates a value is not empty. This allows the system to make sure that a 
+ * required field has not been field with useless data.
  * 
  * @author César de la Cal <cesar@magic3w.com>
  */
-class MinLengthValidationRule implements ValidationRule
+class RegexValidationRule implements ValidationRule
 {
+	
 	/**
-	 * The minimum length a string passed to this element must have for the test
-	 * to be passed successfully.
-	 * @var int
+	 * The regular expression we wish to test the content of the validator against.
+	 *
+	 * @var string
 	 */
-	private $length;
+	private $regex;
 	
 	/**
 	 * A message the validation error generated by this object should carry to give
@@ -26,12 +29,13 @@ class MinLengthValidationRule implements ValidationRule
 	/**
 	 * Additional information given to the user in case the validation did not 
 	 * succeed. This message can hold additional infos on how to solve the error.
+	 * 
 	 * @var string
 	 */
 	private $extendedMessage;
 	
-	public function __construct($length, $message, $extendedMessage = '') {
-		$this->length = $length;
+	public function __construct($regex, $message, $extendedMessage = '') {
+		$this->regex = $regex;
 		$this->message = $message;
 		$this->extendedMessage = $extendedMessage;
 	}
@@ -42,13 +46,13 @@ class MinLengthValidationRule implements ValidationRule
 	 * 
 	 * @param mixed $value
 	 * @param mixed $source
-	 * @return \spitfire\validation\ValidationError|boolean
+	 * @return ValidationError|boolean
 	 */
 	public function test($value) {
-		if (strlen($value) < $this->length) {
+		if (preg_match($this->regex, $value) !== 1) {
 			return new ValidationError($this->message, $this->extendedMessage);
 		}
 		return false;
 	}
-
+	
 }
